Split up the kana mapping into multiple files
