+++ src/index.js
@@ -19,3 +19,5 @@ export { default as toKatakana } from './toKatakana';
 // Other utils
 export { default as stripOkurigana } from './stripOkurigana';
 export { default as tokenize } from './tokenize';
+
+export { createKanaToHepburnMap, applyMapping } from './kanaMapping';
