+++ src/toKana.js
@@ -6,9 +6,10 @@ import isCharUpperCase from './utils/isCharUpperCase';
 import hiraganaToKatakana from './utils/hiraganaToKatakana';
 import {
   getRomajiToKanaTree,
-  applyMapping,
   IME_MODE_MAP,
-  USE_OBSOLETE_KANA_MAP } from './kanaMapping';
+  USE_OBSOLETE_KANA_MAP,
+} from './romajiToKanaMap';
+import { applyMapping } from './kanaMappingUtils';
 
 /**
  * Convert [Romaji](https://en.wikipedia.org/wiki/Romaji) to [Kana](https://en.wikipedia.org/wiki/Kana), lowercase text will result in [Hiragana](https://en.wikipedia.org/wiki/Hiragana) and uppercase text will result in [Katakana](https://en.wikipedia.org/wiki/Katakana).
